Add Java blog admin functionality.

+++ SoftwareTechnologies/JavaBlog/blog/src/main/java/softuniBlog/controller/ArticleController.java
@@ -2,6 +2,7 @@ package softuniBlog.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -55,6 +56,15 @@ public class ArticleController {
 
         Article article = this.articleRepository.findOne(id);
 
+        if (!(SecurityContextHolder.getContext().getAuthentication()
+        instanceof AnonymousAuthenticationToken)) {
+            UserDetails principal = (UserDetails) SecurityContextHolder.getContext()
+                    .getAuthentication().getPrincipal();
+
+            User entityuser = this.userRepository.findByEmail(principal.getUsername());
+            model.addAttribute("user", entityuser);
+        }
+
         model.addAttribute("article", article);
         model.addAttribute("view", "article/details");
 
@@ -68,6 +78,10 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article/" + id;
+        }
 
         model.addAttribute("view", "article/edit");
         model.addAttribute("article", article);
@@ -83,6 +97,10 @@ public class ArticleController {
         }
 
         Article article = this.articleRepository.findOne(id);
+
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article/" + id;
+        }
 
         article.setContent(articleBindingModel.getContent());
         article.setTitle(articleBindingModel.getTitle());
@@ -101,10 +119,22 @@ public class ArticleController {
 
         Article article = this.articleRepository.findOne(id);
 
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article/" + id;
+        }
+
         model.addAttribute("article", article);
         model.addAttribute("view", "article/delete");
 
         return "base-layout";
+    }
+
+    private boolean isUserAuthorOrAdmin(Article article){
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+
+        User userEntity = this.userRepository.findByEmail(user.getUsername());
+        return userEntity.isAdmin() || userEntity.isAuthor(article);
     }
 
     @PostMapping("/article/delete/{id}")
@@ -115,6 +145,11 @@ public class ArticleController {
         }
 
         Article article = this.articleRepository.findOne(id);
+
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article/" + id;
+        }
+
         this.articleRepository.delete(article);
 
         return "redirect:/";
